+++ app/scripts/fhir/R4/lformsFHIR.js
@@ -23638,13 +23638,13 @@ function addCommonSDCFns(ns) {
     'minInclusive': '>=',
     'maxInclusive': '<=',
     'value': '=',
-    'not': '!=',
+    'notEqual': '!=',
     '>': 'minExclusive',
     '<': 'maxExclusive',
     '>=': 'minInclusive',
     '<=': 'maxInclusive',
     '=': 'value',
-    '!=': 'not',
+    '!=': 'notEqual',
     'exists': 'exists'
   };
   /**