+++ bower-dist/lforms.js
@@ -7842,7 +7842,7 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
           // the key is one of the keys in the answers.
           case this._CONSTANTS.DATA_TYPE.CNE:
           case this._CONSTANTS.DATA_TYPE.CWE:
-            var triggerValue = trigger.hasOwnProperty('value') ? trigger.value : trigger.hasOwnProperty('not') ? trigger.not : null;
+            var triggerValue = trigger.hasOwnProperty('value') ? trigger.value : trigger.hasOwnProperty('notEqual') ? trigger.not : null;
             var answerValues = Array.isArray(currentValue) ? currentValue : [currentValue];
             var isEqual = false;
 
@@ -7865,7 +7865,7 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
               if (isEqual) {
                 action = true;
               }
-            } else if (trigger.hasOwnProperty('not')) {
+            } else if (trigger.hasOwnProperty('notEqual')) {
               if (!isEqual) {
                 action = true;
               }
@@ -7884,8 +7884,8 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
               if (trigger["value"] === numCurrentValue) {
                 action = true;
               }
-            } else if (trigger.hasOwnProperty('not')) {
-              if (trigger["not"] != numCurrentValue) {
+            } else if (trigger.hasOwnProperty('notEqual')) {
+              if (trigger["notEqual"] != numCurrentValue) {
                 action = true;
               }
             } // the skip logic rule has a range
@@ -7908,8 +7908,8 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
               if (trigger["value"] === currentValue) {
                 action = true;
               }
-            } else if (trigger.hasOwnProperty('not')) {
-              if (trigger["not"] != currentValue) {
+            } else if (trigger.hasOwnProperty('notEqual')) {
+              if (trigger["notEqual"] != currentValue) {
                 action = true;
               }
             }
@@ -7917,8 +7917,8 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
             break;
         } // end case
 
-      } // no answer and 'not' has a value
-      else if (trigger.hasOwnProperty('not') && trigger.not !== undefined && trigger.not !== null && trigger.not !== "") {
+      } // no answer and 'notEqual' has a value
+      else if (trigger.hasOwnProperty('notEqual') && trigger.notEqual !== undefined && trigger.notEqual !== null && trigger.notEqual !== "") {
           action = true;
         }